Refactor fetchWithRedirect
