+++ src/hooks/api/user.js
@@ -1,9 +1,12 @@
-import { fetchWithRedirect } from '../../utils/fetchWithRedirect';
+import { fetchFromApi } from '../../utils/fetchFromApi';
 import { useAuthenticatedQuery } from './useAuthenticatedQuery';
 
 const getMe = async ({ headers }) => {
-  const res = await fetchWithRedirect('/user', {
-    headers,
+  const res = await fetchFromApi({
+    path: '/user',
+    options: {
+      headers,
+    },
   });
   return await res.json();
 };
@@ -16,8 +19,11 @@ const useGetUser = (configuration = {}) =>
   });
 
 const getPermissions = async ({ headers }) => {
-  const res = await fetchWithRedirect('/user/permissions/', {
-    headers,
+  const res = await fetchFromApi({
+    path: '/user/permissions/',
+    options: {
+      headers,
+    },
   });
   return await res.json();
 };
@@ -30,8 +36,11 @@ const useGetPermissions = (configuration = {}) =>
   });
 
 const getRoles = async ({ headers }) => {
-  const res = await fetchWithRedirect('/user/roles/', {
-    headers,
+  const res = await fetchFromApi({
+    path: '/user/roles/',
+    options: {
+      headers,
+    },
   });
   return await res.json();
 };
