+++ src/utils/fetchFromApi.js
@@ -1,0 +1,19 @@
+import getConfig from 'next/config';
+
+const Errors = {
+  UNAUTHORIZED: 'unauthorized',
+};
+
+const fetchFromApi = async ({ path, searchParams = {}, options = {} }) => {
+  const { publicRuntimeConfig } = getConfig();
+  const url = new URL(`${publicRuntimeConfig.apiUrl}${path}`);
+  url.search = new URLSearchParams(searchParams);
+  const response = await fetch(url, options);
+
+  if (response.status === 401 || response.status === 403) {
+    throw Error(Errors.UNAUTHORIZED);
+  }
+  return response;
+};
+
+export { fetchFromApi, Errors };
